+++ packages/node/src/ethereum/api.ethereum.test.ts
@@ -347,14 +347,17 @@ describe('Api.ethereum', () => {
       transactions: [],
     });
 
-    jest
-      .spyOn((ethApi as any).client, 'getLogs')
-      .mockResolvedValueOnce([
-        { blockHash: mockIncorrectBlockHash, transactionHash: 'tx1' },
-      ]);
+    jest.spyOn((ethApi as any).client, 'getLogs').mockResolvedValueOnce([
+      {
+        blockNumber: '0x1831a96',
+        blockHash: mockIncorrectBlockHash,
+        transactionHash: 'tx1',
+        logIndex: '0x0',
+      },
+    ]);
 
     await expect(ethApi.fetchBlock(mockBlockNumber)).rejects.toThrow(
-      `Log BlockHash does not match block: ${mockBlockNumber}`,
+      `Log BlockHash does not match block: 72194336, blockHash mockBlockHash. Log 0 got block 25369238 blockHash mockIncorrectBlockHash. Please check with rpc provider`,
     );
   });