+++ packages/node/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Improve log hash assertion message (#269)
+
 ## [3.11.0] - 2024-03-20
 ### Changed
 - Update `@subql/node-core` with fixes and optimisations